Refactor project structure and added code comments 🦍
- Project Structure:
Made sure classes are ordered in data and ui packages. Fragments will go in the ui directory and data like the Voertuig model or the API controller go in the data package
