--- app/src/main/java/com/nathan/taxibedrijf_on_the_road/DetailsFragment.kt
@@ -1,11 +1,0 @@
-package com.nathan.taxibedrijf_on_the_road
-
-import android.os.Bundle
-import android.view.View
-import androidx.fragment.app.Fragment
-
-class DetailsFragment : Fragment(R.layout.fragment_details) {
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-}
