--- app/src/main/java/com/nathan/taxibedrijf_on_the_road/Voertuig.kt
@@ -1,8 +1,0 @@
-package com.nathan.taxibedrijf_on_the_road
-
-class Voertuig (var kenteken: String, var voertuigsoort: String){
-
-    override fun toString(): String {
-        return "Kenteken = '$kenteken', Voertuigsoort = '$voertuigsoort'"
-    }
-}
