+++ app/src/main/java/com/nathan/taxibedrijf_on_the_road/data/remote/APIController.kt
@@ -1,4 +1,4 @@
-package com.nathan.taxibedrijf_on_the_road
+package com.nathan.taxibedrijf_on_the_road.data.remote
 
 import android.content.Context
 import android.widget.Toast
@@ -8,6 +8,7 @@ import com.android.volley.toolbox.StringRequest
 import com.android.volley.toolbox.Volley
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
+import com.nathan.taxibedrijf_on_the_road.data.model.Voertuig
 
 class APIController(var context : Context) {
     fun getData(callback: (ArrayList<Voertuig>) -> Unit){
@@ -21,19 +22,21 @@ class APIController(var context : Context) {
         //Vervolgens is er een StringRequest variabele aangemaakt, die een GET functie uitvoert. Dit betekent dat er data wordt opgehaald. Ook de API url is nodig.
         val request = StringRequest(
             Request.Method.GET, url,
+            //variabele response doet iets als er een response is van de API
             {response->
                 //Als je een specifieke array in de JSON file nodig hebt gebruik je:
-                //var jArray = JSONObject(response.toString()).getJSONArray()
+                //var jArray = JSONObject(response.toString()).getJSONArray(*)
 
                 //Maakt een Arraylist van type voertuig objecten.
                 val arrayVoertuig = object: TypeToken<ArrayList<Voertuig>>(){}.type
                 //Maakt een variabele dat de zojuist gemaakte Arraylist is, gevuld met van Json omgezette gson resultaten.
                 val voertuigen : ArrayList<Voertuig> = gson.fromJson(response.toString(), arrayVoertuig)
+                //Geeft de ArrayList terug.
                 callback(voertuigen)
             },
-            //Zo niet, dat geven we een error.
+            //Als er geen response is, dan geven we een error.
             {error->
-                //Dit doen we door Toast te gebruiken, over de gehele applicatie, met als text "Geen Reactie!" en de duur is lang. .show() laat de Toast gelijk zien.
+                //Dit doen we door Toast te gebruiken, met als text "Geen Reactie!" en de duur is lang. .show() laat de Toast gelijk zien.
                 Toast.makeText(context, "Geen reactie!", Toast.LENGTH_LONG).show()
             })
         // Vervolgens moeten we nog wel de GET request in de RequestQueue zetten.
