+++ app/src/main/java/com/nathan/taxibedrijf_on_the_road/ui/profile/ProfileFragment.kt
@@ -1,10 +1,12 @@
-package com.nathan.taxibedrijf_on_the_road
+package com.nathan.taxibedrijf_on_the_road.ui.profile
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
+import com.nathan.taxibedrijf_on_the_road.R
 
 class ProfileFragment : Fragment(R.layout.fragment_profile) {
+    // Functie die nodig is om Fragment inhoud te schrijven.
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
     }
